use wikimedia-ui-base as a base truth
Adds a build script which extracts a tokens.json, as compatible with
style-dictionary, from wikimedia-ui-base and adds those tokens as the
"base truth" to the configuration of style-dictionary.

Also organizes/renames the local token files according to the
agreed-upon terminology (global, alias, component-level).

`global-delta.json` is still quite verbose, containing all the
information it had previously, but showcases the reuse of WMF's
"wmui-color-base0" as "wmui.color.base0.value".

https://amzn.github.io/style-dictionary/#/config?id=configjson